+++ Engine/Engine/Lenn.cpp
@@ -83,7 +83,7 @@ void Lenn::levelUp()
 	stats.resist +=5;
 	stats.speed += 4;
 	stats.xp = 0;
-	xpToLevel += (float) xpToLevel * 0.3;
+	xpToLevel += (int)((float) xpToLevel * 0.3f);
 	if(stats.level == 20)
 		xpToLevel = 0;
 	}
